# test-tools/swagger-src/backend_swagger20.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "1.0.14 (20180124)",
    "title" : "BOCHK FinTech Hackathon 2018 API (Backend)",
    "contact" : {
      "url" : "http://www.bochkhackathon.com/",
      "email" : "it_innovation_lab@bochk.com"
    }
  },
  "host" : "bocapis.mybluemix.net",
  "basePath" : "/api",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/accounts/{account_no}/transactions" : {
      "get" : {
        "tags" : [ "Account Operations" ],
        "summary" : "List account transactions",
        "parameters" : [ {
          "name" : "ibm-app-user",
          "in" : "header",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Status 200",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/account_transaction"
              }
            }
          },
          "404" : {
            "description" : "Status 404"
          }
        }
      },
      "parameters" : [ {
        "name" : "account_no",
        "in" : "path",
        "required" : true,
        "type" : "string"
      } ]
    },
    "/accounts/{account_no}/money-transfer" : {
      "post" : {
        "tags" : [ "Account Operations" ],
        "summary" : "Transfer money (intra-bank)",
        "consumes" : [ ],
        "parameters" : [ {
          "name" : "ibm-app-user",
          "in" : "header",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "body",
          "in" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/money_transfer_request"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Status 200",
            "schema" : {
              "$ref" : "#/definitions/money_transfer_response"
            }
          },
          "400" : {
            "description" : "Status 400"
          }
        }
      },
      "parameters" : [ {
        "name" : "account_no",
        "in" : "path",
        "required" : true,
        "type" : "string"
      } ]
    },
...
  },
  "securityDefinitions" : {
    "Basic authentication" : {
      "type" : "basic"
    }
  },
  "definitions" : {
    "money_transfer_request" : {
      "type" : "object",
      "required" : [ "amount", "currency", "dst_account_no" ],
      "properties" : {
        "currency" : {
          "$ref" : "#/definitions/currency"
        },
        "amount" : {
          "type" : "number",
          "description" : "Amount to transfer (in dollar)"
        },
        "dst_account_no" : {
          "type" : "string",
          "description" : "Account to receive money"
        },
        "remark" : {
          "type" : "string",
          "description" : "Free text remark"
        }
      }
    },
...
    "account_transaction" : {
      "type" : "object",
      "required" : [ "account_no", "after_balance", "amount", "currency", "datetime", "txn_type" ],
      "properties" : {
        "datetime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "txn_type" : {
          "type" : "string",
          "enum" : [ "FOREX", "TRANSFER", "LOAN", "STOCK" ]
        },
        "account_no" : {
          "type" : "string"
        },
        "account_no2" : {
          "type" : "string"
        },
        "fps_phone_no" : {
          "type" : "string"
        },
        "fps_email" : {
          "type" : "string"
        },
        "fps_id" : {
          "type" : "string"
        },
        "currency" : {
          "$ref" : "#/definitions/currency"
        },
        "currency2" : {
          "$ref" : "#/definitions/currency"
        },
        "amount" : {
          "type" : "number"
        },
        "amount2" : {
          "type" : "number"
        },
        "after_balance" : {
          "type" : "number"
        },
        "after_balance2" : {
          "type" : "number"
        },
        "stock_code" : {
          "type" : "string"
        },
        "remark" : {
          "type" : "string"
        }
      }
    },
...
    "currency" : {
      "type" : "string",
      "enum" : [ "HKD", "USD", "CNY", "AUD", "CAD", "CHF", "EUR", "GBP", "JPY", "NZD", "SGD", "THB", "BND", "ZAR", "DKK" ]
    },
...
    "money_transfer_response" : {
      "type" : "object",
      "required" : [ "dst_account_name", "src_balance_after" ],
      "properties" : {
        "src_balance_after" : {
          "type" : "number"
        },
        "dst_account_name" : {
          "type" : "string"
        }
      }
    },
...
  }
}